+++ trackingnumbers.go
@@ -74,12 +74,12 @@ type PackageTrack struct {
 	Currency                   *string  `json:"currency,omitempty"`
 	TotalWeight                *int     `json:"totalWeight,omitempty"`
 	DistributionModality       *int     `json:"distributionModality,omitempty"`
-	PackagingLength            *float32 `json:"packagingLength,omitempty"`
-	PackagingWidth             *float32 `json:"packagingWidth,omitempty"`
-	PackagingHeight            *float32 `json:"packagingHeight,omitempty"`
+	PackagingLength            *float64 `json:"packagingLength,omitempty"`
+	PackagingWidth             *float64 `json:"packagingWidth,omitempty"`
+	PackagingHeight            *float64 `json:"packagingHeight,omitempty"`
 	TaxPaymentMethod           *string  `json:"taxPaymentMethod,omitempty"`
-	FreightPaidValue           *float32 `json:"freightPaidValue,omitempty"`
-	InsurancePaidValue         *float32 `json:"insurancePaidValue,omitempty"`
+	FreightPaidValue           *float64 `json:"freightPaidValue,omitempty"`
+	InsurancePaidValue         *float64 `json:"insurancePaidValue,omitempty"`
 	Items                      []Item   `json:"items,omitempty"`
 }
